Fix overflow issue at cast from float to int
When performing the projection of octant corners into image coordinates,
I faced an edge case were the image coordinate was in the factor of e9.
As afterwards we cast to int, this was causing an overflow, leading to a
negative value. This was problematic since this was used for the
bounding boxes coordinates, leading to the bb_min being larger than
bb_max, since the casting was posterior than selecting the min and max.
This was causing to segfault.

The fix: clamp the max and min values on the image space prior to the
casting.