Remove space for empty structs
In C++ all objects have a size of at least 1, even empty structs. So for
example, the TSDF data without colour or ID would be expected to be
2*sizeof(float)=8 but is instead 2*sizeof(float)+1+1+padding=12.

The [[no_unique_address]] attribute was introduced in C++20 to solve
this issue by marking data members with potentially zero size. Since
unknown attributes are ignored by compilers it is safe to add it even
though we target C++17.

Fortunately, both GCC 9.4.0 and clang 10.0.0 honour the attribute even
in C++17 mode, resulting in a memory usage reduction of 4 bytes per
voxel.