Fix Map::interpField() usage
This fix should have been applied as part of commit a5e80f5 "Combine
Map::getFieldInterp() overloads". Due to the change of arguments
introduced in that commit, the scale_tt variable was passed to the
desired_scale argument of Map::getFieldInterp() instead of the intended
returned_scale argument. This resulted in the scale render always
showing scale 0.